Use NIOAsyncTestingChannel for testing
